Update bbfmm.c
Fixed a bug in malloc for large memory allocation

+++ bbfmm.c
@@ -796,7 +796,7 @@ void ComputeKernelCheb(double *Kweights, int n, kernel_t
   int n3 = n*n*n;            // n3 = n^3
   int dofn3_s = dof.s * n3;
   int dofn3_f = dof.f * n3;
-  int dof2n6 = dofn3_s * dofn3_f; // Total size
+  size_t dof2n6 = dofn3_s * dofn3_f; // Total size
   int Sigma_size;
   int2 cutoff;
